fix: 게시물 상세 페이지에서 데이터 로드되지 않으면 레이아웃 보이지 않게
